requirements.txt
opencv-python
pyzbar
Pillow
tkinter
pyperclip
opencv-python-headless